Adds Jinja template for managing upload description (#13)
Issue: https://codeberg.org/c2s/clips2share/issues/4

Adds a Jinja template for managing the upload description making it easier to support multiple templates. Currently the BBCode is printed to the terminal but in the future it could write this to a file. Icons are embedded into the Jinja template.

Reviewed-on: https://codeberg.org/c2s/clips2share/pulls/13
Co-authored-by: c4ktusc0wboy <philmslim@gmail.com>
Co-committed-by: c4ktusc0wboy <philmslim@gmail.com>